Complete actions first version
Completed a first version for custom actions (insert, remove and show).

In order to achieve this we had to modify following things...

- nlu.yml
we have merge all the insert intents in a unique intent "insert". This because there were some issues of missclassification between insert and delete intent.

- domain.yml
to update the intents and add a new actions named "action_reset_user": to reset username slot, so you can switch to another user by identification intend

- stories.yml and rules.yml (omg this one)
to implements in some way a pattern/form of conversation and activate proprerly the custom actions

#TODO
integrade "deadline" and improve stories/rules and training data to better perform